refactor(stats-extension-compressed): remove build version script
refactor(stats-extension-package-info): remove build version script

refactor(stats-extension-stats-validation-result): remove build version script

refactor(stats-validator-plugin-webpack): remove build version script

refactor(stats-validator-reporter-stats-report): remove build version script

refactor(webpack-stats-extension-compressed): remove build version script

refactor(webpack-stats-extension-package-info): remove build version script
