+++ package.json
@@ -14,8 +14,7 @@
     "lint:eslint": "eslint . --ext .js,.ts --cache",
     "lint:eslint:fix": "npm run lint:eslint -- --fix",
     "build": "lerna run build",
-    "build:versions": "lerna run buildVersionTS && npm run lint:eslint:fix",
-    "prepublishOnly": "npm run build && npm run lint && npm run test"
+    "build:versions": "lerna run buildVersionTS && npm run lint:eslint:fix"
   },
   "repository": "https://github.com/statoscope/statoscope",
   "author": {
