--- packages/stats-extension-compressed/src/version.ts
@@ -1,6 +1,0 @@
-export const name = '@statoscope/stats-extension-compressed';
-export const version = '5.7.0';
-export const description =
-  'Statoscope extension for collecting compressed resource sizes';
-export const author = { name: 'Sergey Melyukov', url: 'https://twitter.com/smelukov' };
-export const homepage = 'https://github.com/statoscope/statoscope/';
