+++ packages/stats-extension-package-info/src/generator.ts
@@ -1,6 +1,7 @@
 import { Extension, ExtensionDescriptor } from '@statoscope/stats/spec/extension';
 import makeResolver, { Resolver } from '@statoscope/helpers/dist/entity-resolver';
-import { name, version, author, homepage, description } from './version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const { name, version, author, homepage, description } = require('../package.json');
 
 export type Format = Extension<Payload>;
 export type InstanceInfo = { version: string };
