--- packages/stats-extension-stats-validation-result/src/version.ts
@@ -1,5 +1,0 @@
-export const name = '@statoscope/stats-extension-stats-validation-result';
-export const version = '5.7.0';
-export const description = 'Statoscope extension that collects stats validation result';
-export const author = { name: 'Sergey Melyukov', url: 'https://twitter.com/smelukov' };
-export const homepage = 'https://github.com/statoscope/statoscope/';
