+++ packages/stats-validator-plugin-webpack/src/rules/diff-deprecated-modules/index.ts
@@ -6,7 +6,8 @@ import { NormalizedModule } from '@statoscope/webpack-model/src/normalize';
 import { WebpackRule } from '../../';
 import { ModuleTarget, normalizeModuleTarget, RawTarget } from '../../helpers';
 import { ExcludeItem, normalizeExclude } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 export type ModuleResultItem = {
   module: NormalizedModule;
