+++ packages/stats-validator-plugin-webpack/src/rules/diff-entry-download-size-limits/index.ts
@@ -7,7 +7,8 @@ import helpers, { Limit, ValueDiff } from '@statoscope/helpers/dist/jora';
 import { APIFnOptions } from '@statoscope/types/types/validation/api';
 import { WebpackRule } from '../../';
 import { ByNameFilterItem, ExcludeItem, normalizeExclude } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 export type Limits = {
   maxSizeDiff?: number | Limit;
