+++ packages/stats-validator-plugin-webpack/src/rules/diff-entry-download-time-limits/index.ts
@@ -13,7 +13,8 @@ import {
   normalizeExclude,
   SerializedExcludeItem,
 } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 type NetworkType = typeof networkListType[number]['name'];
