+++ packages/stats-validator-plugin-webpack/src/rules/entry-download-size-limits/index.ts
@@ -7,7 +7,8 @@ import { APIFnOptions } from '@statoscope/types/types/validation/api';
 import helpers from '@statoscope/helpers/dist/jora';
 import { WebpackRule } from '../../';
 import { ByNameFilterItem, ExcludeItem, normalizeExclude } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 export type Limits = {
   maxSize?: number;
