+++ packages/stats-validator-plugin-webpack/src/rules/entry-download-time-limits/index.ts
@@ -8,7 +8,8 @@ import networkListType from '@statoscope/helpers/dist/network-type-list';
 import { APIFnOptions } from '@statoscope/types/types/validation/api';
 import { WebpackRule } from '../../';
 import { ByNameFilterItem, ExcludeItem, normalizeExclude } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 type NetworkType = typeof networkListType[number]['name'];
