+++ packages/stats-validator-plugin-webpack/src/rules/no-modules-deopts/index.ts
@@ -8,7 +8,8 @@ import { API } from '@statoscope/types/types/validation/api';
 import { RuleDataInput } from '@statoscope/stats-validator/dist/rule';
 import { WebpackRule } from '../../';
 import { ExcludeItem, normalizeExclude } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 export type ModuleResultItem = {
   file: NormalizedFile;
