+++ packages/stats-validator-plugin-webpack/src/rules/restricted-modules/index.ts
@@ -9,7 +9,8 @@ import { RuleDataInput } from '@statoscope/stats-validator/dist/rule';
 import { WebpackRule } from '../../';
 import { ModuleTarget, normalizeModuleTarget, RawTarget } from '../../helpers';
 import { ExcludeItem, normalizeExclude } from '../../limits-helpers';
-import * as version from '../../version';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const version = require('../../../package.json');
 
 export type ModuleResultItem = {
   file: NormalizedFile;
