--- packages/stats-validator-plugin-webpack/src/version.ts
@@ -1,6 +1,0 @@
-export const name = '@statoscope/stats-validator-plugin-webpack';
-export const version = '5.7.0';
-export const description =
-  'Plugin for @statoscope/stats-validator to validate webpack stats';
-export const author = { name: 'Sergey Melyukov', url: 'https://twitter.com/smelukov' };
-export const homepage = 'https://github.com/statoscope/statoscope/';
