--- packages/stats-validator-reporter-stats-report/src/version.ts
@@ -1,5 +1,0 @@
-export const name = '@statoscope/stats-validator-reporter-stats-report';
-export const version = '5.7.0';
-export const description = 'Stats report reporter for @statoscope/stats-validator';
-export const author = { name: 'Sergey Melyukov', url: 'https://twitter.com/smelukov' };
-export const homepage = 'https://github.com/statoscope/statoscope/';
