--- packages/webpack-stats-extension-compressed/src/version.ts
@@ -1,5 +1,0 @@
-export const name = '@statoscope/webpack-stats-extension-compressed';
-export const version = '5.7.0';
-export const description = 'Webpack adapter for @statoscope/stats-extension-compressed';
-export const author = { name: 'Sergey Melyukov', url: 'https://twitter.com/smelukov' };
-export const homepage = 'https://github.com/statoscope/statoscope/';
